+++ CHANGELOG.md
@@ -37,6 +37,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+### Improvements
+
+* [#289](https://github.com/babylonlabs-io/vigilante/pull/289) chore: additional check for getFundingtx
+
 ## v0.22.1
 
 ### Bug Fixes