+++ CHANGELOG.md
@@ -39,6 +39,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 * [#271](https://github.com/babylonlabs-io/vigilante/pull/271) chore: sync the sample config
 
+### Improvements
+
+* [#278](https://github.com/babylonlabs-io/vigilante/pull/278) chore: metrics for censorship detection
+
+
 ## v0.22.1
 
 ### Bug Fixes