+++ CHANGELOG.md
@@ -37,12 +37,11 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
-* [#271](https://github.com/babylonlabs-io/vigilante/pull/271) chore: sync the sample config
-
 ### Improvements
 
 * [#278](https://github.com/babylonlabs-io/vigilante/pull/278) chore: metrics for censorship detection
-
+* [#271](https://github.com/babylonlabs-io/vigilante/pull/271) chore: sync the sample config
+* [#290](https://github.com/babylonlabs-io/vigilante/pull/290) chore: dump config command
 
 ## v0.22.1