+++ CHANGELOG.md
@@ -41,6 +41,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 * [#278](https://github.com/babylonlabs-io/vigilante/pull/278) chore: metrics for censorship detection
 * [#271](https://github.com/babylonlabs-io/vigilante/pull/271) chore: sync the sample config
+* [#289](https://github.com/babylonlabs-io/vigilante/pull/289) chore: additional check for getFundingtx
 * [#290](https://github.com/babylonlabs-io/vigilante/pull/290) chore: dump config command
 
 ## v0.22.1