+++ CHANGELOG.md
@@ -43,6 +43,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 * [#271](https://github.com/babylonlabs-io/vigilante/pull/271) chore: sync the sample config
 * [#289](https://github.com/babylonlabs-io/vigilante/pull/289) chore: additional check for getFundingtx
 * [#290](https://github.com/babylonlabs-io/vigilante/pull/290) chore: dump config command
+* [#291](https://github.com/babylonlabs-io/vigilante/pull/291) chore: bump bbn to rc.9
 
 ## v0.22.1