+++ CHANGELOG.md
@@ -37,6 +37,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## Unreleased
 
+## v0.23.0
+
 ### Improvements
 
 * [#278](https://github.com/babylonlabs-io/vigilante/pull/278) chore: metrics for censorship detection